Merge all EPL code back into a single file
We now always use an explicit `epl-package` structure, which has a
internal field for the underlying package descriptor.  All EPL
operations now determine at runtime, how to access the descriptor.

This makes things a lot easier to handle, and avoids all sorts of fuckup
from the byte compiler (such as expanding macros defined in file A,
though file A is never loaded a runtime).